+++ CHANGELOG.md
@@ -8,6 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- Enriched and ordered get movements, and corrected update movement [PR#48](https://github.com/silvioubaldino/personal-finance/pull/48)
 - Corrected wallets recalculate and add automatic recalculate [PR#47](https://github.com/silvioubaldino/personal-finance/pull/47)
 
 ## Release 05-11-2024